refactor: move global app files into app folder
