+++ packages/game-journal/src/app/App.tsx
